+++ Runtime/Extensions/PhysicsExtension.cs
@@ -1,5 +1,9 @@
 using System;
 using System.Collections.Generic;
+
+#if UNITY_EDITOR
+using UnityEditor;
+#endif
 using UnityEngine;
 
 namespace OcUtility
@@ -9,26 +13,27 @@ namespace OcUtility
         static bool _initialized;
         static Queue<RaycastHit[]> _rayHitBuffer;
         static Queue<Collider[]> _overlapBuffer;
-        
+
         static int raycastBudget;
         static int overlapBudget;
+
         [RuntimeInitializeOnLoadMethod]
         static void Init()
         {
-            if(_initialized) return;
+            if (_initialized) return;
             _initialized = true;
-            
+
             raycastBudget = OcUtilitySettings.Instance == null ? 32 : OcUtilitySettings.Instance.RaycastBudget;
             overlapBudget = OcUtilitySettings.Instance == null ? 32 : OcUtilitySettings.Instance.OverlapBudget;
 
             _rayHitBuffer = new Queue<RaycastHit[]>();
 
-            
+
             for (int i = 0; i < 20; i++)
             {
                 _rayHitBuffer.Enqueue(new RaycastHit[raycastBudget]);
             }
-            
+
             _overlapBuffer = new Queue<Collider[]>();
             for (int i = 0; i < 20; i++)
             {
@@ -46,7 +51,7 @@ namespace OcUtility
             _initialized = false;
             _rayHitBuffer = null;
             Application.quitting -= Release;
-        }  
+        }
 #endif
 
         public static RaycastHit[] GetRaycastHitBuffer()
@@ -55,6 +60,7 @@ namespace OcUtility
 
             return _rayHitBuffer.Dequeue();
         }
+
         public static Collider[] GetOverlapBuffer()
         {
             if (_overlapBuffer.Count == 0) return new Collider[overlapBudget];
@@ -64,12 +70,13 @@ namespace OcUtility
 
         public static void ReturnBuffer(RaycastHit[] buffer)
         {
-            if(buffer.Length < raycastBudget) Array.Resize(ref buffer, raycastBudget);
+            if (buffer.Length < raycastBudget) Array.Resize(ref buffer, raycastBudget);
             _rayHitBuffer.Enqueue(buffer);
         }
+
         public static void ReturnBuffer(Collider[] buffer)
         {
-            if(buffer.Length < overlapBudget) Array.Resize(ref buffer, overlapBudget);
+            if (buffer.Length < overlapBudget) Array.Resize(ref buffer, overlapBudget);
             _overlapBuffer.Enqueue(buffer);
         }
 
@@ -95,7 +102,7 @@ namespace OcUtility
         {
             if (resolution < 1) resolution = 1;
             var points = new Vector3[resolution + 2];
-            
+
             var g = -gravity;
 
             var lowestTimeValue = max_time_y() / (resolution + 2);
@@ -121,5 +128,148 @@ namespace OcUtility
                 return t;
             }
         }
+
+
+        static float ScaledCapsuleRadius(CapsuleCollider capsule)
+        {
+            var scale = capsule.transform.lossyScale;
+            float max;
+            switch (capsule.direction)
+            {
+                case 0: // X-Axis.
+                    max = scale.y > scale.z ? scale.y : scale.z;
+                    break;
+                case 1: // Y-Axis.
+                    max = scale.x > scale.z ? scale.x : scale.z;
+                    break;
+                case 2: // Z-Axis.
+                    max = scale.x > scale.y ? scale.x : scale.y;
+                    break;
+                default: goto case 1;
+            }
+
+            return max * capsule.radius;
+        }
+
+        static float ScaledCapsuleHeight(CapsuleCollider capsule)
+        {
+            var scale = capsule.transform.lossyScale;
+            float height;
+            switch (capsule.direction)
+            {
+                case 0: // X-Axis.
+                    height = capsule.height * scale.x;
+                    break;
+                case 1: // Y-Axis.
+                    height = capsule.height * scale.y;
+                    break;
+                case 2: // Z-Axis.
+                    height = capsule.height * scale.z;
+                    break;
+                default: goto case 1;
+            }
+
+            var scaledRadius = ScaledCapsuleRadius(capsule);
+            if (height < scaledRadius * 2f) height = scaledRadius * 2f;
+            return height;
+        }
+
+        public static void ToWorldSpaceCapsule(this CapsuleCollider capsule, out Vector3 point0, out Vector3 point1,
+            out float radius)
+        {
+            var center = capsule.transform.TransformPoint(capsule.center);
+            radius = 0f;
+            float height = 0f;
+            Vector3 lossyScale = capsule.transform.lossyScale.abs();
+            Vector3 dir = Vector3.zero;
+
+            switch (capsule.direction)
+            {
+                case 0: // x
+                    radius = Mathf.Max(lossyScale.y, lossyScale.z) * capsule.radius;
+                    height = lossyScale.x * capsule.height;
+                    dir = capsule.transform.TransformDirection(Vector3.right);
+                    break;
+                case 1: // y
+                    radius = Mathf.Max(lossyScale.x, lossyScale.z) * capsule.radius;
+                    height = lossyScale.y * capsule.height;
+                    dir = capsule.transform.TransformDirection(Vector3.up);
+                    break;
+                case 2: // z
+                    radius = Mathf.Max(lossyScale.x, lossyScale.y) * capsule.radius;
+                    height = lossyScale.z * capsule.height;
+                    dir = capsule.transform.TransformDirection(Vector3.forward);
+                    break;
+            }
+
+            if (height < radius * 2f)
+            {
+                dir = Vector3.zero;
+            }
+
+            point0 = center + dir * (height * 0.5f - radius);
+            point1 = center - dir * (height * 0.5f - radius);
+        }
+
+        public static Vector3 DirectionAxis(this CapsuleCollider capsule)
+        {
+            return capsule.direction switch
+            {
+                0 => Vector3.right,
+                1 => Vector3.up,
+                2 => Vector3.forward,
+                _ => Vector3.up
+            };
+        }
+
+#if UNITY_EDITOR
+        [MenuItem("CONTEXT/BoxCollider/바운드에 맞게 확장")]
+        static void ExtendBox(MenuCommand command)
+        {
+            var box = command.context as BoxCollider;
+            var childRenderer = box.GetComponentsInChildren<Renderer>();
+
+            var bounds = box.bounds;
+            foreach (var renderer in childRenderer)
+            {
+                bounds.Encapsulate(renderer.bounds);
+            }
+
+            box.center = bounds.center;
+            box.size = bounds.size;
+        }
+        [MenuItem("CONTEXT/CapsuleCollider/바운드에 맞게 확장")]
+        static void ExtendCapsule(MenuCommand command)
+        {
+            var c = command.context as CapsuleCollider;
+            var childRenderer = c.GetComponentsInChildren<Renderer>();
+
+            var bounds = c.bounds;
+            foreach (var renderer in childRenderer)
+            {
+                bounds.Encapsulate(renderer.bounds);
+            }
+
+            c.center = bounds.center;
+            c.radius = bounds.extents.maxXZ();
+            c.height = bounds.size.y;
+        }
+        
+        [MenuItem("CONTEXT/SphereCollider/바운드에 맞게 확장")]
+        static void ExtendSphere(MenuCommand command)
+        {
+            var c = command.context as SphereCollider;
+            var childRenderer = c.GetComponentsInChildren<Renderer>();
+
+            var bounds = c.bounds;
+            foreach (var renderer in childRenderer)
+            {
+                bounds.Encapsulate(renderer.bounds);
+            }
+
+            c.center = bounds.center;
+            c.radius = bounds.extents.max();
+        }
+#endif
     }
 }