+++ Runtime/Printer.cs
@@ -52,58 +52,69 @@ namespace OcUtility
             }
         }
 
-#if UNITY_EDITOR
+        [Conditional("UNITY_EDITOR"), Conditional("DEVELOPMENT_BUILD")]
         public static void Ray(Vector3 start, Vector3 dir, bool depthTest = false)
         {
             Debug.DrawRay(start, dir, Color.white, 0, depthTest);
         }
+        [Conditional("UNITY_EDITOR"), Conditional("DEVELOPMENT_BUILD")]
         public static void Ray(Vector3 start, Vector3 dir, float duration, bool depthTest)
         {
             Debug.DrawRay(start, dir, Color.white, duration, depthTest);
         }
+        [Conditional("UNITY_EDITOR"), Conditional("DEVELOPMENT_BUILD")]
         public static void Ray(Vector3 start, Vector3 dir, Color color, bool depthTest = false)
         {
             Debug.DrawRay(start, dir, color, 0, depthTest);
         }
+        [Conditional("UNITY_EDITOR"), Conditional("DEVELOPMENT_BUILD")]
         public static void Ray(Vector3 start, Vector3 dir, Color color, float duration, bool depthTest = false)
         {
             Debug.DrawRay(start, dir, color, duration, depthTest);
         }
+        [Conditional("UNITY_EDITOR"), Conditional("DEVELOPMENT_BUILD")]
         public static void Ray(Ray ray, bool depthTest = false)
         {
             Debug.DrawRay(ray.origin, ray.direction, Color.white, 0, depthTest);
         }
+        [Conditional("UNITY_EDITOR"), Conditional("DEVELOPMENT_BUILD")]
         public static void Ray(Ray ray, Color color, bool depthTest = false)
         {
             Debug.DrawRay(ray.origin, ray.direction, color, 0, depthTest);
         }
+        [Conditional("UNITY_EDITOR"), Conditional("DEVELOPMENT_BUILD")]
         public static void Ray(Ray ray, float duration, bool depthTest = false)
         {
             Debug.DrawRay(ray.origin, ray.direction, Color.white, duration, depthTest);
         }
+        [Conditional("UNITY_EDITOR"), Conditional("DEVELOPMENT_BUILD")]
         public static void Ray(Ray ray, Color color, float duration, bool depthTest = false)
         {
             Debug.DrawRay(ray.origin, ray.direction, color, duration, depthTest);
         }
         
-        
+        [Conditional("UNITY_EDITOR"), Conditional("DEVELOPMENT_BUILD")]
         public static void Line(Vector3 start, Vector3 end, bool depthTest = false)
         {
             Debug.DrawLine(start, end, Color.white, 0, depthTest);
         }
+        [Conditional("UNITY_EDITOR"), Conditional("DEVELOPMENT_BUILD")]
         public static void Line(Vector3 start, Vector3 end, float duration, bool depthTest)
         {
             Debug.DrawLine(start, end, Color.white, duration, depthTest);
         }
+        [Conditional("UNITY_EDITOR"), Conditional("DEVELOPMENT_BUILD")]
         public static void Line(Vector3 start, Vector3 end, Color color, bool depthTest = false)
         {
             Debug.DrawLine(start, end, color, 0, depthTest);
         }
+        [Conditional("UNITY_EDITOR"), Conditional("DEVELOPMENT_BUILD")]
         public static void Line(Vector3 start, Vector3 end, Color color, float duration, bool depthTest = false)
         {
             Debug.DrawLine(start, end, color, duration, depthTest);
         }
 
+        [Conditional("UNITY_EDITOR"), Conditional("DEVELOPMENT_BUILD")]
         public static void DrawDonut(Vector3 center, Vector3 normal, Vector3 from, 
             float angle, float minRadius, float maxRadius, Color color)
         {
@@ -146,17 +157,21 @@ namespace OcUtility
             GL.PopMatrix();
         }
 
+        [Conditional("UNITY_EDITOR"), Conditional("DEVELOPMENT_BUILD")]
         public static void DrawDonut(Vector3 center, Vector3 normal, Vector3 from, 
             float angle, Vector2 range, Color color)
         {
             DrawDonut(center, normal, from, angle, range.x, range.y, color);
         }
+        
+        [Conditional("UNITY_EDITOR"), Conditional("DEVELOPMENT_BUILD")]
         public static void DrawDonut(Vector3 center, Vector3 from, 
             float angle, Vector2 range, Color color)
         {
             DrawDonut(center, Vector3.up, from, angle, range.x, range.y, color);
         }
 
+        [Conditional("UNITY_EDITOR"), Conditional("DEVELOPMENT_BUILD")]
         public static void DrawDonut(Vector3 center, Vector3 normal, 
             float centerAngle, float angle, float minRadius, float maxRadius, Color color)
         {
@@ -164,12 +179,14 @@ namespace OcUtility
             DrawDonut(center, normal, from, angle, minRadius, maxRadius, color);
         }
 
+        [Conditional("UNITY_EDITOR"), Conditional("DEVELOPMENT_BUILD")]
         public static void DrawDonut(Vector3 center, Vector3 normal, 
             float centerAngle, float angle, Vector2 range, Color color)
         {
             var from = Quaternion.AngleAxis(centerAngle, normal) * Vector3.forward;
             DrawDonut(center, normal, from, angle, range, color);
         }
+        [Conditional("UNITY_EDITOR"), Conditional("DEVELOPMENT_BUILD")]
         public static void DrawDonut(Vector3 center, float centerAngle, 
             float angle, Vector2 range, Color color)
         {
@@ -177,6 +194,7 @@ namespace OcUtility
             DrawDonut(center, Vector3.up, from, angle, range, color);
         }
 
+        [Conditional("UNITY_EDITOR"), Conditional("DEVELOPMENT_BUILD")]
         public static void DrawCross(Vector3 position, float size, Color color)
         {
             CreateGizmoMaterial();
@@ -198,6 +216,7 @@ namespace OcUtility
             GL.PopMatrix();
         }
         
+        
         static void CreateGizmoMaterial()
         {
             if (giz_Mat) return;
@@ -213,6 +232,10 @@ namespace OcUtility
             primitive_Mat = Resources.Load<Material>("e_mat_transparent");
         }
 
+
+#if UNITY_EDITOR
+        // 반환형이 void가 아니라서 직접 전처리기로 감싸줌. 성능이 많이 드는 메서드니까
+        // 차라리 빌드과정에서 발견되면 제외시키는게 빌드에 포함시키는것보다 나음.
         public static GameObject CreatePrimitive(
             PrimitiveType type, 
             Vector3 pos,
@@ -234,7 +257,7 @@ namespace OcUtility
             if (duration > 0) wait.time(duration, () => Object.Destroy(gao));
             
             return gao;
-        }
+        }  
 #endif